requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-ckeditor==6.7.0
redis==5.0.1
googletrans==3.1.0a0
pytest==7.4.4
pytest-django==4.7.0
flake8==7.0.0
gunicorn==21.2.0
drf-yasg==1.21.7